broadband-modem-mbim: fix warnings with -Wtype-limits
  mm-broadband-modem-mbim.c: In function ‘basic_connect_notification_signal_state’:
  /usr/include/glib-2.0/glib/gmacros.h:811:63: error: comparison of unsigned expression < 0 is always false [-Werror=type-limits]
    811 | ine CLAMP(x, low, high)  (((x) > (high)) ? (high) : (((x) < (low)) ? (low) : (x)))
        |                                                           ^

  mm-broadband-modem-mbim.c:2918:19: note: in expansion of macro ‘CLAMP’
   2918 |         quality = CLAMP (rssi == 99 ? 0 : rssi, 0, 31) * 100 / 31;
        |                   ^~~~~

+++ src/mm-broadband-modem-mbim.c
@@ -2920,7 +2920,7 @@ basic_connect_notification_signal_state (MMBroadbandModemMbim *self,
         guint32 quality;
 
         /* Normalize the quality. 99 means unknown, we default it to 0 */
-        quality = CLAMP (rssi == 99 ? 0 : rssi, 0, 31) * 100 / 31;
+        quality = CLAMP (rssi == 99 ? 0 : (gint)rssi, 0, 31) * 100 / 31;
 
         mm_dbg ("Signal state indication: %u --> %u%%", rssi, quality);
         mm_iface_modem_update_signal_quality (MM_IFACE_MODEM (self), quality);
